pyproject.toml
[project]
name = "thebuilders"
version = "0.1.0"
description = ""
authors = [
    {name = "Jared Diaz",email = "jtdiaz@umich.edu"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "testcontainers (>=4.13.2,<5.0.0)",
    "pymysql (>=1.0.2,<2.0.0)",
    "uvicorn (>=0.24.0,<1.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi (>=0.120.1,<0.121.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "anthropic (>=0.72.0,<0.73.0)",
    "google-genai (>=1.51.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "minio (>=7.2.18,<8.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "ultralytics (>=8.3.0,<9.0.0)",
    "torch (>=2.5.0,<3.0.0)",
    "torchvision (>=0.20.0,<1.0.0)",
]
packages = [{ include = "*", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
